fix: buffer now use replace=True due to performance issue

+++ flowrl/config/online/mujoco_config.py
@@ -28,7 +28,7 @@ class Config:
     frame_skip: int
     frame_stack: int
     num_train_envs: int
-    utd: int
+    utd: float
     batch_size: int
     buffer_size: int
     lap_alpha: float
